+++ BitFaster.Caching/ITimePolicy.cs
@@ -8,7 +8,7 @@ namespace BitFaster.Caching
     public interface ITimePolicy
     {
         /// <summary>
-        /// Gets the time to live for items in the cache.
+        /// Gets the time to expire for items in the cache.
         /// </summary>
         TimeSpan TimeToLive { get; }
 
@@ -27,9 +27,9 @@ namespace BitFaster.Caching
         /// Gets the time to live for an item in the cache.
         /// </summary>
         /// <param name="key">The key of the item.</param>
-        /// <param name="timeToLive">If the key exists, the time to live for the item with the specified key.</param>
+        /// <param name="timeToExpire">If the key exists, the time to live for the item with the specified key.</param>
         /// <returns>True if the key exists, otherwise false.</returns>
-        bool TryGetTimeToLive<K>(K key, out TimeSpan timeToLive);
+        bool TryGetTimeToExpire<K>(K key, out TimeSpan timeToExpire);
 
         /// <summary>
         /// Remove all expired items from the cache.
