+++ BitFaster.Caching/Lru/Builder/AsyncConcurrentLruBuilder.cs
@@ -16,7 +16,7 @@ namespace BitFaster.Caching.Lru.Builder
         /// <summary>
         /// Evict after a variable duration specified by an IExpiry instance.
         /// </summary>
-        /// <param name="expiry">The expiry that determines item time to live.</param>
+        /// <param name="expiry">The expiry that determines item time to expire.</param>
         /// <returns>A AsyncConcurrentLruBuilder</returns>
         public AsyncConcurrentLruBuilder<K, V> WithExpiry(IExpiryCalculator<K, V> expiry)
         {
